sqlite: upgrade to SQLite 3.14.0
Downloaded from http://sqlite.com/2016/sqlite-amalgamation-3140000.zip

  $ sha256sum sqlite-amalgamation-3140000.zip
  32843740eaaa4d9ac567d6babe18099ef6f38c6731718c2d37b26d381c629ec8  sqlite-amalgamation-3140000.zip

dist/orig contains the stock sqlite3 code, as packaged in the ZIP file above.

dist contains a copy of dist/orig, but with the Android.patch file applied.
Please see Android.patch for a list of differences between stock and
Android.

The Android specific patch applied cleanly and was regenerated.

This is a routine maintenance release of SQLite that includes a number of new features and performance enhancements as well as fixes for some obscure bugs.  See the changelog (http://www.sqlite.org/releaselog/3_14.html) for details.

Change-Id: I85ee07a6f8e693dfd88ff343fae932ee6be5771d
Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>